Destructuring to files
