+++ src/pages/Trends/LabsTrendsPage.module.scss
@@ -1,0 +1,19 @@
+.wrapper {
+  width: 100%;
+  margin-bottom: 2rem;
+}
+
+.header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.title {
+  margin-right: 0.5rem !important;
+}
+
+.icon {
+  margin-right: 8px;
+  margin-top: -1px;
+}
