+++ src/pages/Trends/TrendTable.module.scss
@@ -15,15 +15,6 @@
   padding: 0;
 }
 
-.change {
-  margin-left: 18px;
-}
-
-.triangle {
-  vertical-align: middle;
-  margin-right: 5px;
-}
-
 .table:global(.ReactTable) {
   & {
     border: none;
