+++ src/pages/Trends/TrendsPage.module.scss
@@ -5,7 +5,6 @@
   .header {
     display: flex;
     align-items: center;
-    justify-content: space-between;
 
     h1 {
       margin-right: 0.5rem !important;
@@ -22,7 +21,8 @@
   width: 100%;
   background-color: rgba($white, 0.8);
   box-shadow: 0 1px 2px rgba(21, 24, 31, 0.02),
-    0 2px 10px rgba(21, 24, 31, 0.04), 0 4px 24px rgba(21, 24, 31, 0.06);
+    0 2px 10px rgba(21, 24, 31, 0.04),
+ 0 4px 24px rgba(21, 24, 31, 0.06);
   margin: 0;
 
   :global(.night-mode) & {
@@ -46,23 +46,3 @@
   }
 }
 /* stylelint-enable */
-
-.tables {
-  display: flex;
-  justify-content: space-between;
-
-  > * {
-    flex: 1;
-    margin-right: 15px;
-
-    &:last-child {
-      flex: 3;
-      margin-right: 0;
-    }
-  }
-}
-
-.icon {
-  margin-right: 8px;
-  margin-top: -1px;
-}
